Architectural Pattern and Design Pattern
For the Architectural Pattern, I chosed the "Model-View-Controller (MVC) Pattern" and for the Design Pattern, I only chosed one Design Pattern for now, so it is the "Observer Design Pattern". If you think that something is incorrect and need changes or something new needs to be added feel free to shoot me a message.